+++ src/bundler.js
@@ -1,34 +1,18 @@
-import { dirname } from 'node:path';
-import { join as joinPath } from 'node:path';
-import { normalize as normalizePath } from 'node:path';
 import { rollup } from 'rollup';
 import convertCJS from '@rollup/plugin-commonjs';
-import dts from 'rollup-plugin-dts';
 import terser from '@rollup/plugin-terser';
-import virtual from '@rollup/plugin-virtual';
 import json from '@rollup/plugin-json';
 import babel from '@rollup/plugin-babel';
 import preset from '@babel/preset-env';
 import typescript from '@rollup/plugin-typescript';
-import ts from 'typescript';
+import bundleTypes from './bundler/bundleTypes.js';
 import generateBanner from './generateBanner.js';
 import { node as nodeTarget } from './targets.js';
-
-/**
- * @type {import('globby').globby}
- */
-let globby;
 
 async function bundler( {
 	onWarn,
 	packageInfo
 } ) {
-	if ( !globby ) {
-		const globbyModule = await import( 'globby' );
-		// eslint-disable-next-line require-atomic-updates
-		globby = globbyModule.globby;
-	}
-
 	await Promise.all( bundleChunks( packageInfo, onWarn ) );
 }
 
@@ -130,103 +114,6 @@ function getRollupOutputConfig( outputPath, banner, format = 'esm' ) {
 		file: outputPath,
 		exports: 'auto'
 	};
-}
-
-async function bundleTypes( {
-	project,
-	sourceFile,
-	outputFile,
-	onWarn = () => {}
-} = {} ) {
-	project = normalizePath( project );
-
-	const tsFiles = await globby( 'src/**/*.ts', {
-		absolute: true,
-		cwd: project
-	} );
-	const compilerOptions = {
-		declaration: true,
-		emitDeclarationOnly: true
-	};
-	const emittedFiles = {};
-
-	// Just to be sure…
-	delete compilerOptions.declarationDir;
-
-	const host = ts.createCompilerHost( compilerOptions );
-	host.writeFile = ( filePath, contents ) => {
-		const relativeFilePath = getRelativePath( filePath );
-
-		emittedFiles[ relativeFilePath ] = contents;
-	};
-
-	// Prepare and emit the d.ts files
-	const program = ts.createProgram( tsFiles, compilerOptions, host );
-	program.emit();
-
-	const input = getOriginalDTsFilePath();
-	const rollupConfig = {
-		input,
-		plugins: [
-			{
-				resolveId: ( imported, importer ) => {
-					// Skip the main file.
-					if ( !importer ) {
-						return null;
-					}
-
-					const importerDir = dirname( importer );
-					const jsExtensionRegex = /\.(m|c)?js$/;
-
-					imported = joinPath( importerDir, imported );
-
-					// We need full file path, with extension here.
-					// Due to that we need to:
-					// 1. Remove JS extension (in ESM-based projects
-					//    TS tends to add it).
-					// 2. Add the .d.ts extension.
-					if ( !imported.endsWith( '.d.ts' ) ) {
-						imported = `${ imported.replace( jsExtensionRegex, '' ) }.d.ts`;
-					}
-
-					return imported;
-				}
-			},
-
-			virtual( emittedFiles ),
-
-			dts()
-		],
-		onwarn: onWarn
-	};
-	const outputConfig = {
-		file: outputFile,
-		format: 'es'
-	};
-	const bundle = await rollup( rollupConfig );
-
-	await bundle.write( outputConfig );
-
-	function getOriginalDTsFilePath() {
-		// We need the relative path to the .d.ts file. So:
-		// 1. Get the relative path via getRelativePath().
-		// 2. Replace the .ts extension with the .d.ts one.
-		const originalFilePath = getRelativePath( sourceFile ).replace( /\.ts$/, '.d.ts' );
-
-		return originalFilePath;
-	}
-
-	function getRelativePath( filePath ) {
-		// We need the relative path to the .d.ts file. So:
-		// 1. Normalize the filePath (just to be sure).
-		// 2. Remove the project path.
-		// 3. Remove the leading slash/backslash.
-		const relativeFilePath = normalizePath( filePath ).
-			replace( project, '' ).
-			replace( /^[/\\]/, '' );
-
-		return relativeFilePath;
-	}
 }
 
 export default bundler;